Fix some issues with user and coupon routes, contollers and services

+++ back/app/Http/Controllers/CouponsController.php
@@ -18,10 +18,15 @@ class CouponsController
     //////////////////////////////////////////
 
     public function create(Request $request) {
+      $validAccess = accessValidator($request, 0, false);
+      if (!$validAccess) {
+        return response()->json(['error' => true, 'data' => 'Unauthorized'], 403);
+      }
+
       $validated = couponDataValidator($request);
       
       if (!$validated['error']) {
-        $result = $this->couponsService->register($validated['data']);
+        $result = $this->couponsService->create($validated['data']);
       }
       else {
         $result = ['error'=> true, 'data' => $validated['data'], 'code' => 422];
@@ -31,21 +36,19 @@ class CouponsController
     }
     //////////////////////////////////////////
 
-    public function assign($couponId, $id) {
-      $result = $this->couponsService->assign($couponId, $id);
-    
-      return response()->json(['error' => $result['error'], 'data' => $result['data']], $result['code']);
-    }
-    //////////////////////////////////////////
-
-    public function generate() {
-      $result = $this->couponsService->generate();
+    public function expire() {
+      $result = $this->couponsService->expire();
     
       return response()->json(['error' => $result['error'], 'data' => $result['data']], $result['code']);
     }
     //////////////////////////////////////////
 
     public function update($id, Request $request) {
+      $validAccess = accessValidator($request, 0, false);
+      if (!$validAccess) {
+        return response()->json(['error' => true, 'data' => 'Unauthorized'], 403);
+      }
+
       $validated = couponDataValidator($request);
       
       if (!$validated['error']) {
@@ -59,21 +62,36 @@ class CouponsController
     }
     //////////////////////////////////////////
 
-    public function delete($id) {
+    public function delete($id, Request $request) {
+      $validAccess = accessValidator($request, 0, false);
+      if (!$validAccess) {
+        return response()->json(['error' => true, 'data' => 'Unauthorized'], 403);
+      }
+
       $result = $this->couponsService->delete($id);
     
       return response()->json(['error' => $result['error'], 'data' => $result['data']], $result['code']);
     }
     //////////////////////////////////////////
 
-    public function getAll() {
+    public function getAll(Request $request) {
+      $validAccess = accessValidator($request, 0, false);
+      if (!$validAccess) {
+        return response()->json(['error' => true, 'data' => 'Unauthorized'], 403);
+      }
+
       $coupons = $this->couponsService->getAll();
 
       return response()->json(['error' => false, 'data' => $coupons], 200);
     }
     //////////////////////////////////////////
 
-    public function filter($status) {
+    public function filter($status, Request $request) {
+      $validAccess = accessValidator($request, 0, false);
+      if (!($status == 'valid' || $validAccess)) {
+        return response()->json(['error' => true, 'data' => 'Unauthorized'], 403);
+      }
+
       $validated = couponStatusValidator($status);
       
       if (!$validated['error']) {
@@ -89,13 +107,6 @@ class CouponsController
   
     public function find($id) {
       $result = $this->couponsService->find($id);
-
-      return response()->json(['error' => $result['error'], 'data' => $result['data']], $result['code']);
-    }
-    //////////////////////////////////////////
-    
-    public function getUsers($id) {
-      $result = $this->couponsService->getUsers($id);
 
       return response()->json(['error' => $result['error'], 'data' => $result['data']], $result['code']);
     }
